[refactor]: reorganize parser services and update import paths
